+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/AuthTests.java
@@ -787,7 +787,7 @@ public class AuthTests extends StandardIntegrationTest {
         try {
             user1.retrieveInfoForUser(SyncCredentials.IdentityProvider.USERNAME_PASSWORD, username2);
             fail("It should not be possible to lookup a user using non admin token");
-        } catch (IllegalArgumentException expected) {
+        } catch (IllegalArgumentException ignored) {
         }
     }
 
@@ -821,6 +821,122 @@ public class AuthTests extends StandardIntegrationTest {
             @Override
             public void onError(ObjectServerError error) {
                 fail(error.getErrorMessage());
+            }
+        });
+    }
+
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void requestPasswordResetAsync() {
+        String email = "foo@bar.baz";
+        UserFactory.createUser(email).logOut();
+
+        // Currently no easy way to see if we actually get an email.
+        // Just verify that the network request can complete successfully.
+        SyncUser.requestPasswordResetAsync(email, Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                looperThread.testComplete();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                fail(error.toString());
+            }
+        });
+    }
+
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void requestResetPassword_unknownEmail() {
+        SyncUser.requestPasswordResetAsync("unknown@realm.io", Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                // Server will respond with SUCCESS if the email is incorrect (for security reasons).
+                looperThread.testComplete();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                fail(error.toString());
+            }
+        });
+    }
+
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void completeResetPassword_invalidToken() {
+        SyncUser.completePasswordResetAsync("invalidToken","newPassword", Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                fail();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                assertEquals(ErrorCode.ACCESS_DENIED, error.getErrorCode());
+                looperThread.testComplete();
+            }
+        });
+    }
+
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void requestEmailConfirmation() {
+        String email = "foo@bar.baz";
+        UserFactory.createUser(email).logOut();
+
+        // Currently no easy way to see if we actually get an email.
+        // Just verify that the network request can complete successfully.
+        SyncUser.requestEmailConfirmationAsync(email, Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                looperThread.testComplete();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                fail(error.toString());
+            }
+        });
+    }
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void requestEmailConfirmation_invalidEmail() {
+        SyncUser.requestEmailConfirmationAsync("unknown@realm.io", Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                // Server will respond with SUCCESS if the email is incorrect (for security reasons).
+                looperThread.testComplete();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                fail(error.toString());
+            }
+        });
+    }
+
+
+    @Test
+    @RunTestInLooperThread
+    @Ignore("Depends on https://github.com/realm/realm-java/pull/5909")
+    public void confirmEmail_invalidToken() {
+        SyncUser.confirmEmailAsync("invalidToken", Constants.AUTH_URL, new SyncUser.Callback<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                fail();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                assertEquals(ErrorCode.ACCESS_DENIED, error.getErrorCode());
+                looperThread.testComplete();
             }
         });
     }